+++ bin/env.bak.js
@@ -2,3 +2,4 @@ process.env.TOKEN = '';
 process.env.CLIENT_ID = '';
 process.env.GUILD_ID = '';
 process.env.TIMEOUT_EXEMPT = `[""]`;
+process.env.PLAYERS = `[]`;