+++ bin/initialize.js
@@ -3,13 +3,20 @@ require('./env');
 const { REST } = require('@discordjs/rest');
 const { Routes } = require('discord-api-types/v9');
 
-const commands = [{
-  name: 'attack',
-  type: 2
-}, {
-  name: 'heal',
-  type: 2
-}]; 
+const commands = [
+  {
+    name: 'attack',
+    type: 2
+  }, 
+  {
+    name: 'heal',
+    type: 2
+  },
+  {
+    name: 'swing',
+    description: 'Takes a swing at a random member of the room.'
+  }
+]; 
 
 const rest = new REST({ version: '9' }).setToken(process.env.TOKEN);