+++ index.js
@@ -5,6 +5,7 @@ const { Client, Intents } = require('discord.js');
 const client = new Client({ intents: [Intents.FLAGS.GUILDS] });
 const attackCommand = require('./bin/commands/AttackCommand');
 const healCommand = require('./bin/commands/HealCommand');
+const swingCommand = require('./bin/commands/SwingCommand');
 
 client.on('ready', () => {
   console.log(`Logged in as ${client.user.tag}!`);
@@ -18,6 +19,14 @@ client.on('interactionCreate', async interaction => {
 
     if(interaction.commandName === 'heal') {
       healCommand.heal(interaction);
+    }
+
+    if(interaction.commandName === 'swing') {
+      client.channels.fetch(interaction.channelId)
+        .then(channel => {
+          swingCommand.swing(interaction, channel);
+        })
+      
     }
 
 });