+++ CHANGES.md
@@ -7,3 +7,7 @@
 - Added Dice and Players modules.
 - Added a heal command.
 - Changed attack command to utilize new modules.
+- Added a check for death when a player is attacked.
+- Added checks for bot targeting.
+- All sorts of updates to the commands.
+- Added a /swing command that randomly attacks a member of the channel.